chore: rename namespaces and classes
