+++ src/AliceNovel/App.xaml
@@ -1,8 +1,8 @@
 ﻿<?xml version = "1.0" encoding = "UTF-8" ?>
 <Application xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:local="clr-namespace:Alice_Novel"
-             x:Class="Alice_Novel.App">
+             xmlns:local="clr-namespace:AliceNovel"
+             x:Class="AliceNovel.App">
     <Application.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
