+++ src/AliceNovel/AppShell.xaml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <Shell
-    x:Class="Alice_Novel.AppShell"
+    x:Class="AliceNovel.AppShell"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:local="clr-namespace:Alice_Novel"
+    xmlns:local="clr-namespace:AliceNovel"
     Shell.FlyoutBehavior="Disabled">
 
     <TabBar>
