+++ src/AliceNovel/Platforms/Android/MainActivity.cs
@@ -2,7 +2,7 @@
 using Android.Content.PM;
 using Android.OS;
 
-namespace Alice_Novel
+namespace AliceNovel
 {
     [Activity(Theme = "@style/Maui.SplashTheme", MainLauncher = true, ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation | ConfigChanges.UiMode | ConfigChanges.ScreenLayout | ConfigChanges.SmallestScreenSize | ConfigChanges.Density)]
     public class MainActivity : MauiAppCompatActivity
