+++ src/AliceNovel/SettingPage.xaml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Alice_Novel.SettingPage">
+             x:Class="AliceNovel.SettingPage">
 
 </ContentPage>
